fix: add isin pushdown, move to separate module
